pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "interject"
version = "0.1.0"
description = "Ambient discovery and research engine — scans arXiv, HN, GitHub, Anthropic docs for new capabilities"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MK", email = "mistakeknot@vibeguider.org" }
]
keywords = ["research", "discovery", "recommendation", "mcp-server"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "sentence-transformers>=2.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
interject-mcp = "interject.server:main"
interject-scan = "interject.scanner:main"

[project.urls]
Homepage = "https://github.com/mistakeknot/interject"
Repository = "https://github.com/mistakeknot/interject"

[tool.hatch.build.targets.wheel]
packages = ["src/interject"]
